+++ README.md
@@ -18,3 +18,7 @@ Ready to run in production? Please [check our deployment guides](https://hexdocs
   * Docs: https://hexdocs.pm/phoenix
   * Mailing list: http://groups.google.com/group/phoenix-talk
   * Source: https://github.com/phoenixframework/phoenix
+
+## Mix tasks
+
+  * Download the latest Dokku database with `mix pull_dokku_database` (dokku git remote needs to be set)
